Type of field id changed to long.

+++ training-solutions/src/main/java/activitytracker/Activity.java
@@ -5,19 +5,19 @@ import activity.ActivityType;
 import java.time.LocalDateTime;
 
 public class Activity {
-    private int id;
+    private long id;
     private LocalDateTime startTime;
     private String description;
     private activity.ActivityType type;
 
-    public Activity(int id, LocalDateTime startTime, String description, ActivityType type) {
+    public Activity(long id, LocalDateTime startTime, String description, ActivityType type) {
         this.id = id;
         this.startTime = startTime;
         this.description = description;
         this.type = type;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
